Finish off environment test movement and merging
